Removed logarithmic scaling of Windows MultiMedia volume control.
Fixed microphone volume control for XP as it does not have a "master" volume control. This has the unfortunate side affect of only the microphone has a volume control and if the user selects line in or something it does not work. But as 99.9% of the time it is a microphone, we live with the limitation.

git-svn-id: https://svn.code.sf.net/p/opalvoip/code/ptlib/branches/v2_10@28254 023b2edf-31b2-4de3-b41e-bca80c47788f